ExtensiveLodashReplacementPlugin : Use `nmf.getResolver` to find modules. (#34441)
* Use `enhanced-resolve` to find modules.

Switch to webpack's `enhanced-resolve` package to find modules.
This saves us from running the algorithm ourselves and parsing JSON.

The version matching algorithm has been changed, since we no longer have
access to the requested version range. It now checks whether the major
versions match and the root `lodash-es` is the same or newer than the
imported package.

* Use nmf.getResolver instead of enhanced-resolver.

* Fix potential race conditions and reorganize code.

* More review changes.

* More review changes.

* Remove `package-lock.json` for plugin.
