+++ packages/webpack-extensive-lodash-replacement-plugin/index.js
@@ -1,405 +1,121 @@
 /* eslint-disable import/no-extraneous-dependencies */
 /* eslint-disable import/no-nodejs-modules */
 const path = require( 'path' );
-const findPackageJson = require( 'find-package-json' );
-const fs = require( 'fs' );
 const semver = require( 'semver' );
 
-// List of all known lodash module names.
-// Used for normalizing names to avoid code duplication.
-// Note: this list should be kept up-to-date with new lodash versions.
-const LODASH_MODULE_NAMES = [
-	'add',
-	'after',
-	'ary',
-	'assign',
-	'assignIn',
-	'assignInWith',
-	'assignWith',
-	'at',
-	'attempt',
-	'before',
-	'bind',
-	'bindAll',
-	'bindKey',
-	'camelCase',
-	'capitalize',
-	'castArray',
-	'ceil',
-	'chain',
-	'chunk',
-	'clamp',
-	'clone',
-	'cloneDeep',
-	'cloneDeepWith',
-	'cloneWith',
-	'commit',
-	'compact',
-	'concat',
-	'cond',
-	'conforms',
-	'conformsTo',
-	'constant',
-	'countBy',
-	'create',
-	'curry',
-	'curryRight',
-	'debounce',
-	'deburr',
-	'defaultTo',
-	'defaults',
-	'defaultsDeep',
-	'defer',
-	'delay',
-	'difference',
-	'differenceBy',
-	'differenceWith',
-	'divide',
-	'drop',
-	'dropRight',
-	'dropRightWhile',
-	'dropWhile',
-	'each',
-	'eachRight',
-	'endsWith',
-	'entries',
-	'entriesIn',
-	'eq',
-	'escape',
-	'escapeRegExp',
-	'every',
-	'extend',
-	'extendWith',
-	'fill',
-	'filter',
-	'find',
-	'findIndex',
-	'findKey',
-	'findLast',
-	'findLastIndex',
-	'findLastKey',
-	'first',
-	'flatMap',
-	'flatMapDeep',
-	'flatMapDepth',
-	'flatten',
-	'flattenDeep',
-	'flattenDepth',
-	'flip',
-	'floor',
-	'flow',
-	'flowRight',
-	'forEach',
-	'forEachRight',
-	'forIn',
-	'forInRight',
-	'forOwn',
-	'forOwnRight',
-	'fromPairs',
-	'functions',
-	'functionsIn',
-	'get',
-	'groupBy',
-	'gt',
-	'gte',
-	'has',
-	'hasIn',
-	'head',
-	'identity',
-	'inRange',
-	'includes',
-	'indexOf',
-	'initial',
-	'intersection',
-	'intersectionBy',
-	'intersectionWith',
-	'invert',
-	'invertBy',
-	'invoke',
-	'invokeMap',
-	'isArguments',
-	'isArray',
-	'isArrayBuffer',
-	'isArrayLike',
-	'isArrayLikeObject',
-	'isBoolean',
-	'isBuffer',
-	'isDate',
-	'isElement',
-	'isEmpty',
-	'isEqual',
-	'isEqualWith',
-	'isError',
-	'isFinite',
-	'isFunction',
-	'isInteger',
-	'isLength',
-	'isMap',
-	'isMatch',
-	'isMatchWith',
-	'isNaN',
-	'isNative',
-	'isNil',
-	'isNull',
-	'isNumber',
-	'isObject',
-	'isObjectLike',
-	'isPlainObject',
-	'isRegExp',
-	'isSafeInteger',
-	'isSet',
-	'isString',
-	'isSymbol',
-	'isTypedArray',
-	'isUndefined',
-	'isWeakMap',
-	'isWeakSet',
-	'iteratee',
-	'join',
-	'kebabCase',
-	'keyBy',
-	'keys',
-	'keysIn',
-	'last',
-	'lastIndexOf',
-	'lodash',
-	'lowerCase',
-	'lowerFirst',
-	'lt',
-	'lte',
-	'map',
-	'mapKeys',
-	'mapValues',
-	'matches',
-	'matchesProperty',
-	'max',
-	'maxBy',
-	'mean',
-	'meanBy',
-	'memoize',
-	'merge',
-	'mergeWith',
-	'method',
-	'methodOf',
-	'min',
-	'minBy',
-	'mixin',
-	'multiply',
-	'negate',
-	'next',
-	'noop',
-	'now',
-	'nth',
-	'nthArg',
-	'omit',
-	'omitBy',
-	'once',
-	'orderBy',
-	'over',
-	'overArgs',
-	'overEvery',
-	'overSome',
-	'pad',
-	'padEnd',
-	'padStart',
-	'parseInt',
-	'partial',
-	'partialRight',
-	'partition',
-	'pick',
-	'pickBy',
-	'plant',
-	'property',
-	'propertyOf',
-	'pull',
-	'pullAll',
-	'pullAllBy',
-	'pullAllWith',
-	'pullAt',
-	'random',
-	'range',
-	'rangeRight',
-	'rearg',
-	'reduce',
-	'reduceRight',
-	'reject',
-	'remove',
-	'repeat',
-	'replace',
-	'rest',
-	'result',
-	'reverse',
-	'round',
-	'sample',
-	'sampleSize',
-	'set',
-	'setWith',
-	'shuffle',
-	'size',
-	'slice',
-	'snakeCase',
-	'some',
-	'sortBy',
-	'sortedIndex',
-	'sortedIndexBy',
-	'sortedIndexOf',
-	'sortedLastIndex',
-	'sortedLastIndexBy',
-	'sortedLastIndexOf',
-	'sortedUniq',
-	'sortedUniqBy',
-	'split',
-	'spread',
-	'startCase',
-	'startsWith',
-	'stubArray',
-	'stubFalse',
-	'stubObject',
-	'stubString',
-	'stubTrue',
-	'subtract',
-	'sum',
-	'sumBy',
-	'tail',
-	'take',
-	'takeRight',
-	'takeRightWhile',
-	'takeWhile',
-	'tap',
-	'template',
-	'templateSettings',
-	'throttle',
-	'thru',
-	'times',
-	'toArray',
-	'toFinite',
-	'toInteger',
-	'toIterator',
-	'toJSON',
-	'toLength',
-	'toLower',
-	'toNumber',
-	'toPairs',
-	'toPairsIn',
-	'toPath',
-	'toPlainObject',
-	'toSafeInteger',
-	'toString',
-	'toUpper',
-	'transform',
-	'trim',
-	'trimEnd',
-	'trimStart',
-	'truncate',
-	'unary',
-	'unescape',
-	'union',
-	'unionBy',
-	'unionWith',
-	'uniq',
-	'uniqBy',
-	'uniqWith',
-	'uniqueId',
-	'unset',
-	'unzip',
-	'unzipWith',
-	'update',
-	'updateWith',
-	'upperCase',
-	'upperFirst',
-	'value',
-	'valueOf',
-	'values',
-	'valuesIn',
-	'without',
-	'words',
-	'wrap',
-	'wrapperAt',
-	'wrapperChain',
-	'wrapperCommit',
-	'wrapperLodash',
-	'wrapperNext',
-	'wrapperPlant',
-	'wrapperReverse',
-	'wrapperToIterator',
-	'wrapperValue',
-	'xor',
-	'xorBy',
-	'xorWith',
-	'zip',
-	'zipObject',
-	'zipObjectDeep',
-	'zipWith',
-];
+const LODASH_MODULE_NAMES = require( './module-names' );
 
-function throwError( message, error = null ) {
-	throw new Error(
-		`[ExtensiveLodashReplacementPlugin] ${ message }${ error ? ' Error: ' : '' }${ error }`
+function createError( message, error = null ) {
+	return new Error(
+		`[ExtensiveLodashReplacementPlugin] ${ message }${ error ? ` Error: ${ error }` : '' }`
 	);
 }
 
+function getModuleForPath( moduleResolver, rootPath, packageName ) {
+	return new Promise( ( resolve, reject ) => {
+		moduleResolver.resolve(
+			{},
+			rootPath,
+			packageName,
+			{},
+			( error, resource, resourceResolveData ) => {
+				if ( error ) {
+					reject(
+						createError(
+							`Could not find module ${ packageName } for import on ${ rootPath }.`,
+							error
+						)
+					);
+				}
+
+				resolve( resourceResolveData );
+			}
+		);
+	} );
+}
+
 class ExtensiveLodashReplacementPlugin {
-	constructor( baseFile = './package.json' ) {
-		this.baseFile = path.resolve( baseFile );
-		let data;
+	constructor( { baseDir = '.' } = {} ) {
+		this.baseDir = path.resolve( baseDir );
+	}
+
+	async initBaseLodashData() {
+		let baseLodash, baseLodashVersion;
 
 		try {
-			data = fs.readFileSync( this.baseFile );
+			baseLodash = await getModuleForPath( this.moduleResolver, this.baseDir, 'lodash' );
 		} catch ( error ) {
-			throwError( 'Could not read root package.json.', error );
+			throw createError( 'Could not find root `lodash`.' );
 		}
 
 		try {
-			const packageJson = JSON.parse( data );
-			this.lodashVersion =
-				packageJson && packageJson.dependencies && packageJson.dependencies[ 'lodash-es' ];
+			baseLodashVersion =
+				baseLodash && baseLodash.descriptionFileData && baseLodash.descriptionFileData.version;
 		} catch ( error ) {
-			throwError( 'Could not parse root package.json.', error );
+			throw createError( 'Could not determine root `lodash` version.' );
 		}
 
-		if ( ! this.lodashVersion ) {
-			throwError( 'No lodash-es dependency in root package.json.' );
+		try {
+			const baseLodashES = await getModuleForPath( this.moduleResolver, this.baseDir, 'lodash-es' );
+			this.baseLodashESVersion =
+				baseLodashES &&
+				baseLodashES.descriptionFileData &&
+				baseLodashES.descriptionFileData.version;
+		} catch ( error ) {
+			throw createError( 'Could not find root `lodash-es`.' );
 		}
 
-		if ( ! semver.valid( this.lodashVersion ) ) {
-			throwError( `Invalid root package.json lodash-es version: ${ this.lodashVersion }.` );
+		if ( ! this.baseLodashESVersion ) {
+			throw createError( 'Could not determine root `lodash-es` version.' );
+		}
+
+		if ( baseLodashVersion !== this.baseLodashESVersion ) {
+			throw createError( 'Root `lodash` and `lodash-es` versions do not match.' );
 		}
 	}
 
-	// Figure out the requested Lodash version range for a given import.
-	// It looks at the file with the import and traverses upwards, until it finds a package.json file
-	// with the requested dependency. It returns this range.
-	findRequestedLodashRange( file, packageName ) {
-		const finder = findPackageJson( path.dirname( file ) );
-		let version;
+	// Figure out the version for a given import.
+	// It follows the node resolution algorithm until it finds the package, returning its version.
+	async findRequestedVersion( file, packageName ) {
+		const foundResolveData = await getModuleForPath(
+			this.moduleResolver,
+			path.dirname( file ),
+			packageName
+		);
 
-		try {
-			while ( ! version ) {
-				const found = finder.next();
-
-				if ( found.filename === this.baseFile ) {
-					return this.lodashVersion;
-				}
-
-				const contents = found && found.value;
-				version = contents && contents.dependencies && contents.dependencies[ packageName ];
-			}
-		} catch ( error ) {
-			throwError( `Could not find requested range for import on ${ file }.`, error );
-		}
-
-		return version;
+		return (
+			foundResolveData &&
+			foundResolveData.descriptionFileData &&
+			foundResolveData.descriptionFileData.version
+		);
 	}
 
 	// Figure out if the requested Lodash import can be replaced with global lodash-es.
-	// It takes the importer's lodash version range and the global lodash-es version into account.
-	canBeReplaced( file, packageName ) {
-		const requestedVersion = this.findRequestedLodashRange( file, packageName );
-		return requestedVersion && semver.satisfies( this.lodashVersion, requestedVersion );
+	// It takes the importer's version and the global lodash-es version into account.
+	async canBeReplaced( file, packageName ) {
+		const importVersion = await this.findRequestedVersion( file, packageName );
+		const isVersionMatch =
+			importVersion &&
+			semver.major( this.baseLodashESVersion ) === semver.major( importVersion ) &&
+			semver.gte( this.baseLodashESVersion, importVersion );
+
+		if ( ! isVersionMatch ) {
+			const relativePath = path.relative( this.baseDir, file );
+			// Output compilation warning.
+			this.compilation.warnings.push(
+				new Error(
+					`${ relativePath }\n  ${ packageName } version ${ importVersion } cannot be replaced by lodash-es version ${
+						this.baseLodashESVersion
+					}`
+				)
+			);
+		}
+
+		return isVersionMatch;
 	}
 
 	// Get the modified request
-	getModifiedRequest( result ) {
+	async getModifiedRequest( result ) {
 		const { request } = result;
 
 		if ( ! result.contextInfo || ! result.contextInfo.issuer ) {
@@ -408,21 +124,21 @@ class ExtensiveLodashReplacementPlugin {
 
 		// Replace plain 'lodash' with 'lodash-es'.
 		if ( /^lodash$/.test( request ) ) {
-			if ( this.canBeReplaced( result.contextInfo.issuer, 'lodash' ) ) {
+			if ( await this.canBeReplaced( result.contextInfo.issuer, 'lodash' ) ) {
 				return 'lodash-es';
 			}
 		}
 
 		// Replace 'lodash/foo' with 'lodash-es/foo'.
 		if ( /^lodash\/(.*)$/.test( request ) ) {
-			if ( this.canBeReplaced( result.contextInfo.issuer, 'lodash' ) ) {
+			if ( await this.canBeReplaced( result.contextInfo.issuer, 'lodash' ) ) {
 				return request.replace( 'lodash/', 'lodash-es/' );
 			}
 		}
 
 		// Replace 'lodash.foo' with 'lodash-es/foo'.
 		if ( /^lodash\.(.*)$/.test( request ) ) {
-			if ( this.canBeReplaced( result.contextInfo.issuer, request ) ) {
+			if ( await this.canBeReplaced( result.contextInfo.issuer, request ) ) {
 				const match = /^lodash\.(.*)$/.exec( request );
 				let subModule = match[ 1 ];
 
@@ -442,18 +158,68 @@ class ExtensiveLodashReplacementPlugin {
 		return request;
 	}
 
+	async modifyResult( result ) {
+		if ( ! result || ! result.contextInfo || ! result.contextInfo.issuer ) {
+			return;
+		}
+
+		// Wait for initialization, if it's still running.
+		await this.init;
+
+		const { request } = result;
+
+		// Replace plain 'lodash' with 'lodash-es'.
+		if ( /^lodash$/.test( request ) ) {
+			if ( await this.canBeReplaced( result.contextInfo.issuer, 'lodash' ) ) {
+				result.request = 'lodash-es';
+				return;
+			}
+		}
+
+		// Replace 'lodash/foo' with 'lodash-es/foo'.
+		if ( /^lodash\/(.*)$/.test( request ) ) {
+			if ( await this.canBeReplaced( result.contextInfo.issuer, 'lodash' ) ) {
+				result.request = request.replace( 'lodash/', 'lodash-es/' );
+				return;
+			}
+		}
+
+		// Replace 'lodash.foo' with 'lodash-es/foo'.
+		if ( /^lodash\.(.*)$/.test( request ) ) {
+			if ( await this.canBeReplaced( result.contextInfo.issuer, request ) ) {
+				const match = /^lodash\.(.*)$/.exec( request );
+				let subModule = match[ 1 ];
+
+				// Normalize module names.
+				// This avoids code duplication due to module name case differences
+				// (e.g. 'camelcase' vs 'camelCase').
+				LODASH_MODULE_NAMES.forEach( casedModule => {
+					if ( subModule === casedModule.toLowerCase() ) {
+						subModule = casedModule;
+					}
+				} );
+
+				result.request = `lodash-es/${ subModule }`;
+				return;
+			}
+		}
+
+		return;
+	}
+
 	apply( compiler ) {
+		compiler.hooks.thisCompilation.tap( 'LodashReplacementPlugin', compilation => {
+			this.compilation = compilation;
+		} );
+
 		compiler.hooks.normalModuleFactory.tap( 'LodashReplacementPlugin', nmf => {
-			nmf.hooks.beforeResolve.tap( 'LodashReplacementPlugin', result => {
-				if ( ! result ) return;
-				result.request = this.getModifiedRequest( result );
-				return result;
-			} );
-			nmf.hooks.afterResolve.tap( 'LodashReplacementPlugin', result => {
-				if ( ! result ) return;
-				result.request = this.getModifiedRequest( result );
-				return result;
-			} );
+			this.moduleResolver = this.moduleResolver || nmf.getResolver( 'normal' );
+			this.init = this.init || this.initBaseLodashData();
+
+			nmf.hooks.beforeResolve.tapPromise(
+				'LodashReplacementPlugin',
+				this.modifyResult.bind( this )
+			);
 		} );
 	}
 }
